requirements.txt
google-cloud-aiplatform
langchain
streamlit
langchain-core 
langchain-google-vertexai
python-dotenv
langchain_community